+++ iceoryx_hoofs/include/iceoryx_hoofs/cxx/requires.hpp
@@ -1,0 +1,46 @@
+// Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2022 by Apex.AI Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+#ifndef IOX_HOOFS_CXX_REQUIRES_HPP
+#define IOX_HOOFS_CXX_REQUIRES_HPP
+
+#include "iceoryx_hoofs/platform/platform_correction.hpp"
+
+namespace iox
+{
+namespace cxx
+{
+namespace internal
+{
+void Require(
+    const bool condition, const char* file, const int line, const char* function, const char* conditionString) noexcept;
+} // namespace internal
+
+// implementing C++ Core Guideline, I.6. Prefer Expects
+// see:
+// https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Ri-expects
+#define Expects(condition) internal::Require(condition, __FILE__, __LINE__, __PRETTY_FUNCTION__, #condition)
+
+// implementing C++ Core Guideline, I.8. Prefer Ensures
+// see:
+// https://github.com/isocpp/CppCoreGuidelines/blob/master/CppCoreGuidelines.md#Ri-ensures
+#define Ensures(condition) internal::Require(condition, __FILE__, __LINE__, __PRETTY_FUNCTION__, #condition)
+
+
+} // namespace cxx
+} // namespace iox
+
+#endif